email in stdout in dev base settings in multi-files
